added and modified chat and user controller, serviceimpl, chat and user repo

+++ src/main/java/com/whatsappclone/service/UserService.java
@@ -3,16 +3,16 @@ package com.whatsappclone.service;
 import java.util.List;
 
 import com.whatsappclone.exception.UserException;
-import com.whatsappclone.modal.User;
+import com.whatsappclone.model.User;
 import com.whatsappclone.request.UpdateUserRequest;
 
 public interface UserService {
 
-	public User findUserById(Integer id) throws UserException;
+	public User findUserById(Long id) throws UserException;
 
 	public User findUserProfile(String jwt) throws UserException;
 
-	public User updateUser(Integer userId, UpdateUserRequest req) throws UserException;
+	public User updateUser(Long userId, UpdateUserRequest req) throws UserException;
 
 	public List<User> searchUser(String query);
